refactor(shared-server): split group mutation validation
